requirements.txt
django
djangorestframework
langchain-core
langchain-google-genai
python-dotenv
numpy
dj-database-url